jupyterhub/requirements.txt
setuptools==40.3.0
jupyterhub-kubespawner==0.11.1
jupyterhub-dummyauthenticator
PyMySQL
oauthenticator==0.8.2
python-jose==3.2.0